Reverse partially 64d6502
